created tick variable intialisation function.
