+++ c_code/comms_s.h
@@ -29,6 +29,6 @@ void s_dlrv_packet     (void);
 
 void s_lds_packet  (uint payload);
 void s_bsgn_packet (void);
-void s_sync_packet (void);
+void s_sync_packet (uint key);
 
 #endif
