+++ c_code/comms_w.h
@@ -26,7 +26,7 @@ void w_processBKPQueue      (uint unused0, uint unused1);
 void w_forward_packet  (uint key, uint payload);
 void w_stop_packet     (uint key);
 void w_net_stop_packet (uint key);
-void w_sync_packet     (void);
+void w_sync_packet     (uint key);
 void w_dlrv_packet     (void);
 
 void w_ldsa_packet (uint payload);
