+++ c_code/process_t.c
@@ -28,7 +28,6 @@
 #include "mlp_externs.h"
 
 #include "init_t.h"
-#include "comms_t.h"
 #include "process_t.h"
 #include "activation.h"
 
@@ -221,36 +220,10 @@ void tf_advance_tick (uint unused0, uint unused1)
   fsg_recv = 0;
 #endif
 
-  // initialise thread semaphore,
-  tf_thrds_pend = tf_thrds_init;
+  // prepare to restart tick,
+  tick_init (!SPINN_RESTART);
 
-  // initialise scoreboards,
-  tf_arrived = 0;
-  tf_crit_arrived = 0;
-
-  // initialise previous value,
-  tf_crit_prev = TRUE;
-
-  // if recording all ticks,
-  if (!xcfg.rec_last_tick_only)
-  {
-    if (t_rec_tick_data)
-    {
-      record_tick_data ();
-    }
-
-    if (t_rec_outputs)
-    {
-      record_outputs ();
-    }
-
-    if (t_rec_step_updt)
-    {
-      stage_step++;
-    }
-  }
-
-  // check if done with event
+  // and check if done with event
   if (tick_stop)
   {
     // update event criterion
@@ -292,20 +265,10 @@ void tb_advance_tick (uint unused0, uint unused1)
   bsg_recv = 0;
 #endif
 
-  // initialise thread semaphore,
-  tb_thrds_pend = tb_thrds_init;
+  // prepare to restart tick,
+  tick_init (!SPINN_RESTART);
 
-  // initialise arrival scoreboards,
-  tb_arrived = 0;
-  tb_bsgn_arrived = 0;
-
-  // update pointer to processing unit outputs,
-  tb_procs = 1 - tb_procs;
-
-  // update pointer to received errors,
-  tb_comms = 1 - tb_comms;
-
-  // check if done with BACKPROP phase
+  // and check if done with BACKPROP phase
   if (tick == SPINN_TB_END_TICK)
   {
     // initialise the tick count
@@ -1141,5 +1104,200 @@ void error_cross_entropy (uint inx)
   {
     t_output_deriv[inx] = 0;
   }
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// in the FORWARD phase the convergence criterion may require the simulation to
+// stop before the maximum time is reached. This routine sends a broadcast
+// message to communicate the final decision if the criterion has been reached
+// across all the output groups to all the cores in the simulation
+// ------------------------------------------------------------------------
+void send_stop_crit (void)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "send_stop_crit\n");
+#endif
+
+  // "aggregate" criteria,
+  tf_stop_crit = tf_stop_crit && tf_crit_prev;
+
+  // make stop decision,
+  if (tcfg.is_last_output)
+  {
+    tf_group_crit = tf_stop_crit;
+
+    if (!xcfg.training)
+    {
+      t_test_results.examples_correct += tf_stop_crit && (ev_tick >=min_ticks);
+    }
+
+    tf_stop_crit = (ev_tick >= max_ticks)
+                     || (tick == ncfg.global_max_ticks - 1)
+                     || (tf_stop_crit && (ev_tick >= min_ticks));
+    tick_stop = tf_stop_crit;
+  }
+
+  // FORWARD aggregated criterion - stop decision if last t core
+  uint stop_crit = (tf_stop_crit) ? SPINN_BOOL_ONE : SPINN_BOOL_ZERO;
+  uint pkt_key = tf_stop_key | stop_crit;
+
+#ifdef DEBUG
+  // include tick (for synchronisation checks)
+  pkt_key |= (tick & SPINN_TICK_MASK);
+#endif
+
+  uint use_pl = tcfg.is_last_output ? NO_PAYLOAD : WITH_PAYLOAD;
+
+  while (!spin1_send_mc_packet (pkt_key, 0, use_pl));
+
+#ifdef DEBUG
+  if (tcfg.is_last_output) stp_sent++;
+  else crt_sent++;
+#endif
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// stores the net of the specified unit for the current tick
+// ------------------------------------------------------------------------
+void store_net (uint inx)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "store_net\n");
+#endif
+
+  t_net_history[(tick * tcfg.num_units) + inx] = t_nets[inx];
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// restores the net of the specified unit for the requested tick
+// ------------------------------------------------------------------------
+void restore_net (uint inx, uint tick)
+{
+#ifdef TRACE
+    io_printf (IO_BUF, "restore_net\n");
+#endif
+
+  t_nets[inx] = t_net_history[((tick * tcfg.num_units) + inx)];
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// stores the output of the specified unit for the current tick
+// ------------------------------------------------------------------------
+void store_output (uint inx)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "store_output\n");
+#endif
+
+  t_output_history[(tick * tcfg.num_units) + inx] = t_outputs[inx];
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// restores the output of the specified unit for the requested tick
+// ------------------------------------------------------------------------
+void restore_output (uint inx, uint tick)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "restore_output\n");
+#endif
+
+  t_outputs[inx] = t_output_history[((tick * tcfg.num_units) + inx)];
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// restores the output of the specified unit for the requested tick
+// ------------------------------------------------------------------------
+void restore_outputs (uint tick)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "restore_outputs\n");
+#endif
+
+  for (uint inx = 0; inx < tcfg.num_units; inx++)
+  {
+    t_outputs[inx] = t_output_history[((tick * tcfg.num_units) + inx)];
+  }
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// stores the output derivative of the specified unit for the current tick
+// ------------------------------------------------------------------------
+void store_output_deriv (uint inx)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "store_output_deriv\n");
+#endif
+
+  t_output_deriv_history[(tick * tcfg.num_units) + inx] = t_output_deriv[inx];
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// restores the output derivative of the specified unit for the requested tick
+// ------------------------------------------------------------------------
+void restore_output_deriv (uint inx, uint tick)
+{
+#ifdef TRACE
+  io_printf (IO_BUF, "restore_output_deriv\n");
+#endif
+
+  t_output_deriv[inx] =
+    t_output_deriv_history[(tick * tcfg.num_units) + inx];
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// record outputs - to be picked up by the host
+// ------------------------------------------------------------------------
+void record_outputs (void)
+{
+  // cast outputs to the right size,
+  short_activ_t outputs[tcfg.num_units];
+
+  for (uint i = 0; i < tcfg.num_units; i++)
+  {
+    outputs[i] = (short_activ_t) (t_outputs[i]
+            >> (SPINN_ACTIV_SHIFT - SPINN_SHORT_ACTIV_SHIFT));
+  }
+
+  // record outputs,
+  recording_record(OUTPUTS,
+      (void *) outputs, tcfg.num_units * sizeof (short_activ_t)
+  );
+}
+// ------------------------------------------------------------------------
+
+
+// ------------------------------------------------------------------------
+// record tick data - to be picked up by the host
+// ------------------------------------------------------------------------
+void record_tick_data (void)
+{
+  tick_record_t tick_data;
+
+  // prepare tick data,
+  tick_data.epoch   = epoch;
+  tick_data.example = example_cnt;
+  tick_data.event   = evt;
+  tick_data.tick    = tick;
+
+  // and record it
+  recording_record(TICK_DATA, (void *) &tick_data, sizeof (tick_record_t));
 }
 // ------------------------------------------------------------------------
