+++ c_code/process_t.h
@@ -48,4 +48,17 @@ void max_stop_crit       (uint inx);
 void error_cross_entropy (uint inx);
 void error_squared       (uint inx);
 
+void send_stop_crit (void);
+
+void store_net            (uint inx);
+void restore_net          (uint inx, uint tick);
+void store_output         (uint inx);
+void restore_output       (uint inx, uint tick);
+void restore_outputs      (uint tick);
+void store_output_deriv   (uint inx);
+void restore_output_deriv (uint inx, uint tick);
+
+void record_outputs   (void);
+void record_tick_data (void);
+
 #endif
